chore: Tests clean up safety measure

+++ tests/clean_up.py
@@ -8,7 +8,7 @@ def delete_test_bigquery_datasets():
     """
     datasets = BigQuery().ls()
     datasets_to_delete = [d for d in datasets if d.startswith("test_")]
-    del_fun = lambda x: BigQuery(dataset=x).delete()
+    del_fun = lambda x: BigQuery(dataset=x).delete() if x.startswith("test_") else None
     with ThreadPoolExecutor() as executor:
         executor.map(del_fun, datasets_to_delete)
 
@@ -19,11 +19,11 @@ def delete_test_firestore_collections():
     """
     collections = Firestore().ls()
     collections_to_delete = [c for c in collections if c.startswith("test_")]
-    del_fun = lambda x: Firestore(x).delete()
+    del_fun = lambda x: Firestore(x).delete() if x.startswith("test_") else None
     with ThreadPoolExecutor() as executor:
         executor.map(del_fun, collections_to_delete)
 
 
 if __name__ == "__main__":
-    # delete_test_bigquery_datasets()
+    delete_test_bigquery_datasets()
     delete_test_firestore_collections()
